requirements.txt
alembic==1.7.7
aniso8601==9.0.1
appnope==0.1.2
asn1crypto==1.3.0
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
cffi==1.13.2
click==7.1.2
constantly==15.1.0
cryptography==3.3.2
cssselect==1.1.0
Cython==0.29.28
decorator==4.4.1
demjson==2.2.4
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-Migrate==3.1.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
hyperlink==18.0.0
idna==2.8
importlib-metadata==4.11.3
importlib-resources==5.6.0
incremental==17.5.0
ipdb==0.11
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.3
lxml==4.6.5
Mako==1.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
parsel==1.5.1
parso==0.7.0
passlib==1.7.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
Protego==0.1.16
psycopg2==2.9.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.0.0
Pygments==2.7.4
PyHamcrest==2.0.2
pymongo==3.10.1
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2022.1
queuelib==1.5.0
Scrapy==1.8.1
scrapyrt==0.11.0
service-identity==18.1.0
six==1.14.0
SQLAlchemy==1.4.34
traitlets==4.3.2
Twisted==20.3.0
typing_extensions==4.1.1
w3lib==1.19.0
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.7.0
zope.interface==4.7.1
